Fix scoring bugs the right way
Everything that I said in 5851c81 was complete bullshit. After further
investigation there was no asynchronous behavior going on (I thought it
was odd that it would so I started probing), the clear was taking
multiple frames to process because it would start at the bottom, and as
it deleted frames the next line up would shift downward, getting skipped.
Simply clearing from the top down fixes this.