Fix bug where tetrominos stuck themselves to walls
It is possible to stick a piece to a wall, but it was happening in cases
where it shouldn't. Once a piece has entered lock down, it has 30 frames
until it fully locks. In this time, if moved in any direction, or
rotated, the timer resets. These timer resets are referred to as
"extensions". You get 15 extensions before the piece immediately locks.

BUT, when a piece moves downward, the number of extensions are reset,
UNLESS the piece has been rotated (while in lock down) such that it is
higher on the grid than before. At this point, you get 15 side-to-side
movements or rotations before it locks against the first surface it
touches. So, while you can totally stick a piece to the wall like this
intentionally, it was counting downward motion in with this, and the piece
would stick to a wall while just sliding downwards without any input (having
previously e.g. rotated over an obstacle), causing me to lose.